Improve readability be extracting separate types to their own files
