+++ Sources/Logging/Logger.swift
@@ -12,22 +12,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-#if canImport(Darwin)
-import Darwin
-#elseif os(Windows)
-import CRT
-#elseif canImport(Glibc)
-@preconcurrency import Glibc
-#elseif canImport(Android)
-@preconcurrency import Android
-#elseif canImport(Musl)
-import Musl
-#elseif canImport(WASILibc)
-import WASILibc
-#else
-#error("Unsupported runtime")
-#endif
-
 /// A Logger emits log messages using methods that correspond to a log level.
 ///
 /// `Logger` is a value type with respect to the ``Logger/Level`` and the ``Metadata`` (as well as the immutable `label`
@@ -1078,192 +1062,6 @@ extension Logger {
     }
 }
 
-/// The logging system is a global facility where you can configure the default logging backend implementation.
-///
-/// `LoggingSystem` is set up just once in a given program to set up the desired logging backend implementation.
-/// The default behavior, if you don't define otherwise, sets the ``LogHandler`` to use a ``StreamLogHandler`` that presents its output to `STDOUT`.
-///
-/// You can configure that handler to present the output to `STDERR` instead using the following code:
-///
-/// ```swift
-/// LoggingSystem.bootstrap(StreamLogHandler.standardError)
-/// ```
-///
-/// The default (``StreamLogHandler``) is intended to be a convenience.
-/// For production applications, implement the ``LogHandler`` protocol directly, or use a community-maintained backend.
-public enum LoggingSystem: Sendable {
-    private static let _factory = FactoryBox(
-        { label, _ in StreamLogHandler.standardError(label: label) },
-        violationErrorMesage: "logging system can only be initialized once per process."
-    )
-    private static let _metadataProviderFactory = MetadataProviderBox(
-        nil,
-        violationErrorMesage: "logging system can only be initialized once per process."
-    )
-
-    #if DEBUG
-    private static let _warnOnceBox: WarnOnceBox = WarnOnceBox()
-    #endif
-
-    /// A one-time configuration function that globally selects the implementation for your desired logging backend.
-    ///
-    /// >  Warning:
-    /// > `bootstrap` can be called at maximum once in any given program, calling it more than once will
-    /// > lead to undefined behavior, most likely a crash.
-    ///
-    /// - parameters:
-    ///     - factory: A closure that provides a ``Logger`` label identifier and produces an instance of the ``LogHandler``.
-    @preconcurrency
-    public static func bootstrap(_ factory: @escaping @Sendable (String) -> any LogHandler) {
-        self._factory.replace(
-            { label, _ in
-                factory(label)
-            },
-            validate: true
-        )
-    }
-
-    /// A one-time configuration function that globally selects the implementation for your desired logging backend.
-    ///
-    /// >  Warning:
-    /// > `bootstrap` can be called at maximum once in any given program, calling it more than once will
-    /// > lead to undefined behavior, most likely a crash.
-    ///
-    /// - parameters:
-    ///     - metadataProvider: The `MetadataProvider` used to inject runtime-generated metadata from the execution context.
-    ///     - factory: A closure that provides a ``Logger`` label identifier and produces an instance of the ``LogHandler``.
-    @preconcurrency
-    public static func bootstrap(
-        _ factory: @escaping @Sendable (String, Logger.MetadataProvider?) -> any LogHandler,
-        metadataProvider: Logger.MetadataProvider?
-    ) {
-        self._metadataProviderFactory.replace(metadataProvider, validate: true)
-        self._factory.replace(factory, validate: true)
-    }
-
-    // for our testing we want to allow multiple bootstrapping
-    internal static func bootstrapInternal(_ factory: @escaping @Sendable (String) -> any LogHandler) {
-        self._metadataProviderFactory.replace(nil, validate: false)
-        self._factory.replace(
-            { label, _ in
-                factory(label)
-            },
-            validate: false
-        )
-    }
-
-    // for our testing we want to allow multiple bootstrapping
-    internal static func bootstrapInternal(
-        _ factory: @escaping @Sendable (String, Logger.MetadataProvider?) -> any LogHandler,
-        metadataProvider: Logger.MetadataProvider?
-    ) {
-        self._metadataProviderFactory.replace(metadataProvider, validate: false)
-        self._factory.replace(factory, validate: false)
-    }
-
-    fileprivate static var factory: (String, Logger.MetadataProvider?) -> any LogHandler {
-        { label, metadataProvider in
-            self._factory.underlying(label, metadataProvider)
-        }
-    }
-
-    /// System wide ``Logger/MetadataProvider`` that was configured during the logging system's `bootstrap`.
-    ///
-    /// When creating a ``Logger`` using the plain ``Logger/init(label:)`` initializer, this metadata provider
-    /// will be provided to it.
-    ///
-    /// When using custom log handler factories, make sure to provide the bootstrapped metadata provider to them,
-    /// or the metadata will not be filled in automatically using the provider on log-sites. While using a custom
-    /// factory to avoid using the bootstrapped metadata provider may sometimes be useful, usually it will lead to
-    /// un-expected behavior, so make sure to always propagate it to your handlers.
-    public static var metadataProvider: Logger.MetadataProvider? {
-        self._metadataProviderFactory.underlying
-    }
-
-    #if DEBUG
-    /// Used to warn only once about a specific ``LogHandler`` type when it does not support ``Logger/MetadataProvider``,
-    /// but an attempt was made to set a metadata provider on such handler. In order to avoid flooding the system with
-    /// warnings such warning is only emitted in debug mode, and even then at-most once for a handler type.
-    internal static func warnOnceLogHandlerNotSupportedMetadataProvider<Handler: LogHandler>(
-        _ type: Handler.Type
-    ) -> Bool {
-        self._warnOnceBox.warnOnceLogHandlerNotSupportedMetadataProvider(type: type)
-    }
-    #endif
-
-    /// Protects an object such that it can only be accessed through a Reader-Writer lock.
-    final class RWLockedValueBox<Value: Sendable>: @unchecked Sendable {
-        private let lock = ReadWriteLock()
-        private var storage: Value
-
-        init(initialValue: Value) {
-            self.storage = initialValue
-        }
-
-        func withReadLock<Result>(_ operation: (Value) -> Result) -> Result {
-            self.lock.withReaderLock {
-                operation(self.storage)
-            }
-        }
-
-        func withWriteLock<Result>(_ operation: (inout Value) -> Result) -> Result {
-            self.lock.withWriterLock {
-                operation(&self.storage)
-            }
-        }
-    }
-
-    /// Protects an object by applying the constraints that it can only be accessed through a Reader-Writer lock
-    /// and can only be updated once from the initial value given.
-    private struct ReplaceOnceBox<BoxedType: Sendable> {
-        private struct ReplaceOnce: Sendable {
-            private var initialized = false
-            private var _underlying: BoxedType
-            private let violationErrorMessage: String
-
-            mutating func replaceUnderlying(_ underlying: BoxedType, validate: Bool) {
-                precondition(!validate || !self.initialized, self.violationErrorMessage)
-                self._underlying = underlying
-                self.initialized = true
-            }
-
-            var underlying: BoxedType {
-                self._underlying
-            }
-
-            init(underlying: BoxedType, violationErrorMessage: String) {
-                self._underlying = underlying
-                self.violationErrorMessage = violationErrorMessage
-            }
-        }
-
-        private let storage: RWLockedValueBox<ReplaceOnce>
-
-        init(_ underlying: BoxedType, violationErrorMesage: String) {
-            self.storage = .init(
-                initialValue: ReplaceOnce(
-                    underlying: underlying,
-                    violationErrorMessage: violationErrorMesage
-                )
-            )
-        }
-
-        func replace(_ newUnderlying: BoxedType, validate: Bool) {
-            self.storage.withWriteLock { $0.replaceUnderlying(newUnderlying, validate: validate) }
-        }
-
-        var underlying: BoxedType {
-            self.storage.withReadLock { $0.underlying }
-        }
-    }
-
-    private typealias FactoryBox = ReplaceOnceBox<
-        @Sendable (_ label: String, _ provider: Logger.MetadataProvider?) -> any LogHandler
-    >
-
-    private typealias MetadataProviderBox = ReplaceOnceBox<Logger.MetadataProvider?>
-}
-
 extension Logger {
     /// The type of the metadata storage.
     ///
@@ -1551,687 +1349,6 @@ extension Logger {
     }
 }
 
-/// A pseudo log handler that sends messages to multiple other log handlers.
-///
-/// ### Effective Logger.Level
-///
-/// When first initialized, the multiplex log handlers' log level is automatically set to the minimum of all the
-/// provided log handlers.
-/// This ensures that each of the handlers are able to log at their appropriate level
-/// any log events they might be interested in.
-///
-/// Example:
-/// If log handler `A` is logging at `.debug` level, and log handler `B` is logging at `.info` level, the log level of the constructed
-/// `MultiplexLogHandler([A, B])` is set to `.debug`. This means that this handler will operate on debug messages,
-/// while only logged by the underlying `A` log handler (since `B`'s log level is `.info`
-/// and thus it would not actually log that log message).
-///
-/// If the log level is _set_ on a `Logger` backed by an `MultiplexLogHandler` the log level applies to *all*
-/// underlying log handlers, allowing a logger to still select at what level it wants to log regardless of if the underlying
-/// handler is a multiplex or a normal one. If for some reason one might want to not allow changing a log level of a specific
-/// handler passed into the multiplex log handler, this is possible by wrapping it in a handler which ignores any log level changes.
-///
-/// ### Effective Logger.Metadata
-///
-/// Since a `MultiplexLogHandler` is a combination of multiple log handlers, the handling of metadata can be non-obvious.
-/// For example, the underlying log handlers may have metadata of their own set before they are used to initialize the multiplex log handler.
-///
-/// The multiplex log handler acts purely as proxy and does not make any changes to underlying handler metadata other than
-/// proxying writes that users made on a `Logger` instance backed by this handler.
-///
-/// Setting metadata is always proxied through to _all_ underlying handlers, meaning that if a modification like
-/// `logger[metadataKey: "x"] = "y"` is made, all the underlying log handlers used to create the multiplex handler
-/// observe this change.
-///
-/// Reading metadata from the multiplex log handler MAY need to pick one of conflicting values if the underlying log handlers
-/// were previously initiated with metadata before passing them into the multiplex handler. The multiplex handler uses
-/// the order in which the handlers were passed in during its initialization as a priority indicator - the first handler's
-/// values are more important than the next handlers values, etc.
-///
-/// Example:
-/// If the multiplex log handler was initiated with two handlers like this: `MultiplexLogHandler([handler1, handler2])`.
-/// The handlers each have some already set metadata: `handler1` has metadata values for keys `one` and `all`, and `handler2`
-/// has values for keys `two` and `all`.
-///
-/// A query through the multiplex log handler the key `one` naturally returns `handler1`'s value, and a query for `two`
-/// naturally returns `handler2`'s value.
-/// Querying for the key `all` will return `handler1`'s value, as that handler has a high priority,
-/// as indicated by its earlier position in the initialization, than the second handler.
-/// The same rule applies when querying for the `metadata` property of the multiplex log handler; it constructs `Metadata` uniquing values.
-public struct MultiplexLogHandler: LogHandler {
-    private var handlers: [any LogHandler]
-    private var effectiveLogLevel: Logger.Level
-    /// This metadata provider runs after all metadata providers of the multiplexed handlers.
-    private var _metadataProvider: Logger.MetadataProvider?
-
-    /// Create a multiplex log handler.
-    ///
-    /// - parameters:
-    ///    - handlers: An array of `LogHandler`s, each of which will receive the log messages sent to this `Logger`.
-    ///                The array must not be empty.
-    public init(_ handlers: [any LogHandler]) {
-        assert(!handlers.isEmpty, "MultiplexLogHandler.handlers MUST NOT be empty")
-        self.handlers = handlers
-        self.effectiveLogLevel = handlers.map { $0.logLevel }.min() ?? .trace
-    }
-
-    /// Create a multiplex log handler with the metadata provider you provide.
-    /// - Parameters:
-    ///   - handlers: An array of `LogHandler`s, each of which will receive every log message sent to this `Logger`.
-    ///    The array must not be empty.
-    ///   - metadataProvider: The metadata provider that adds metadata to log messages for this handler.
-    public init(_ handlers: [any LogHandler], metadataProvider: Logger.MetadataProvider?) {
-        assert(!handlers.isEmpty, "MultiplexLogHandler.handlers MUST NOT be empty")
-        self.handlers = handlers
-        self.effectiveLogLevel = handlers.map { $0.logLevel }.min() ?? .trace
-        self._metadataProvider = metadataProvider
-    }
-
-    /// Get or set the log level configured for this `Logger`.
-    ///
-    /// > Note: Changing the log level threshold for a logger only affects the instance of the `Logger` where you change it.
-    /// > It is acceptable for logging backends to have some form of global log level override
-    /// > that affects multiple or even all loggers. This means a change in `logLevel` to one `Logger` might in
-    /// > certain cases have no effect.
-    public var logLevel: Logger.Level {
-        get {
-            self.effectiveLogLevel
-        }
-        set {
-            self.mutatingForEachHandler { $0.logLevel = newValue }
-            self.effectiveLogLevel = newValue
-        }
-    }
-
-    /// The metadata provider.
-    public var metadataProvider: Logger.MetadataProvider? {
-        get {
-            if self.handlers.count == 1 {
-                if let innerHandler = self.handlers.first?.metadataProvider {
-                    if let multiplexHandlerProvider = self._metadataProvider {
-                        return .multiplex([innerHandler, multiplexHandlerProvider])
-                    } else {
-                        return innerHandler
-                    }
-                } else if let multiplexHandlerProvider = self._metadataProvider {
-                    return multiplexHandlerProvider
-                } else {
-                    return nil
-                }
-            } else {
-                var providers: [Logger.MetadataProvider] = []
-                let additionalMetadataProviderCount = (self._metadataProvider != nil ? 1 : 0)
-                providers.reserveCapacity(self.handlers.count + additionalMetadataProviderCount)
-                for handler in self.handlers {
-                    if let provider = handler.metadataProvider {
-                        providers.append(provider)
-                    }
-                }
-                if let multiplexHandlerProvider = self._metadataProvider {
-                    providers.append(multiplexHandlerProvider)
-                }
-                guard !providers.isEmpty else {
-                    return nil
-                }
-                return .multiplex(providers)
-            }
-        }
-        set {
-            self.mutatingForEachHandler { $0.metadataProvider = newValue }
-        }
-    }
-
-    /// Log a message using the log level and source that you provide.
-    ///
-    /// - parameters:
-    ///    - event: The log event containing the level, message, metadata, and source location.
-    public func log(event: LogEvent) {
-        for handler in self.handlers where handler.logLevel <= event.level {
-            handler.log(event: event)
-        }
-    }
-
-    @available(*, deprecated, renamed: "log(event:)")
-    public func log(
-        level: Logger.Level,
-        message: Logger.Message,
-        metadata: Logger.Metadata?,
-        source: String,
-        file: String,
-        function: String,
-        line: UInt
-    ) {
-        self.log(
-            event: LogEvent(
-                level: level,
-                message: message,
-                metadata: metadata,
-                source: source,
-                file: file,
-                function: function,
-                line: line
-            )
-        )
-    }
-
-    /// Get or set the entire metadata storage as a dictionary.
-    public var metadata: Logger.Metadata {
-        get {
-            var effective: Logger.Metadata = [:]
-            // as a rough estimate we assume that the underlying handlers have a similar metadata count,
-            // and we use the first one's current count to estimate how big of a dictionary we need to allocate:
-
-            // !-safe, we always have at least one handler
-            effective.reserveCapacity(self.handlers.first!.metadata.count)
-
-            for handler in self.handlers {
-                effective.merge(handler.metadata, uniquingKeysWith: { _, handlerMetadata in handlerMetadata })
-                if let provider = handler.metadataProvider {
-                    effective.merge(provider.get(), uniquingKeysWith: { _, provided in provided })
-                }
-            }
-            if let provider = self._metadataProvider {
-                effective.merge(provider.get(), uniquingKeysWith: { _, provided in provided })
-            }
-
-            return effective
-        }
-        set {
-            self.mutatingForEachHandler { $0.metadata = newValue }
-        }
-    }
-
-    /// Add, change, or remove a logging metadata item.
-    ///
-    /// > Note: Changing the logging metadata only affects the instance of the `Logger` where you change it.
-    public subscript(metadataKey metadataKey: Logger.Metadata.Key) -> Logger.Metadata.Value? {
-        get {
-            for handler in self.handlers {
-                if let value = handler[metadataKey: metadataKey] {
-                    return value
-                }
-            }
-            return nil
-        }
-        set {
-            self.mutatingForEachHandler { $0[metadataKey: metadataKey] = newValue }
-        }
-    }
-
-    private mutating func mutatingForEachHandler(_ mutator: (inout any LogHandler) -> Void) {
-        for index in self.handlers.indices {
-            mutator(&self.handlers[index])
-        }
-    }
-}
-
-/// A wrapper to facilitate `print`-ing to stderr and stdio that
-/// ensures access to the underlying `FILE` is locked to prevent
-/// cross-thread interleaving of output.
-internal struct StdioOutputStream: TextOutputStream, @unchecked Sendable {
-    internal let flushMode: FlushMode
-    private let _writeBytes: (UnsafeBufferPointer<UInt8>) -> Void
-    private let _flush: () -> Void
-
-    // Using a generic initializer lets Swift infer the concrete file-pointer type
-    // at each call site from the values passed. This avoids a CFilePointer typealias
-    // that cannot vary by Android API level (API 23 exposes FILE as
-    // UnsafeMutablePointer<FILE>; API 24+ makes it OpaquePointer).
-    internal init<F>(
-        file: F,
-        flushMode: FlushMode,
-        lock: ((F) -> Void)?,
-        unlock: ((F) -> Void)?,
-        write: @escaping (UnsafeRawPointer, Int, Int, F) -> Int,
-        flush: @escaping (F) -> Int32
-    ) {
-        self.flushMode = flushMode
-        self._writeBytes = { bytes in
-            lock?(file)
-            defer { unlock?(file) }
-            if let base = bytes.baseAddress, bytes.count > 0 {
-                _ = write(base, 1, bytes.count, file)
-            }
-        }
-        self._flush = { _ = flush(file) }
-    }
-
-    internal func write(_ string: String) {
-        self.contiguousUTF8(string).withContiguousStorageIfAvailable { utf8Bytes in
-            self._writeBytes(utf8Bytes)
-            if case .always = self.flushMode {
-                self.flush()
-            }
-        }!
-    }
-
-    /// Flush the underlying stream.
-    internal func flush() {
-        self._flush()
-    }
-
-    internal func contiguousUTF8(_ string: String) -> String.UTF8View {
-        var contiguousString = string
-        contiguousString.makeContiguousUTF8()
-        return contiguousString.utf8
-    }
-
-    internal static let stderr: StdioOutputStream = {
-        #if canImport(Darwin)
-        return StdioOutputStream(
-            file: Darwin.stderr,
-            flushMode: .always,
-            lock: flockfile,
-            unlock: funlockfile,
-            write: fwrite,
-            flush: fflush
-        )
-        #elseif os(Windows)
-        return StdioOutputStream(
-            file: CRT.stderr,
-            flushMode: .always,
-            lock: _lock_file,
-            unlock: _unlock_file,
-            write: fwrite,
-            flush: fflush
-        )
-        #elseif canImport(Glibc)
-        #if os(FreeBSD) || os(OpenBSD)
-        return StdioOutputStream(
-            file: Glibc.stderr,
-            flushMode: .always,
-            lock: flockfile,
-            unlock: funlockfile,
-            write: fwrite,
-            flush: fflush
-        )
-        #else
-        return StdioOutputStream(
-            file: Glibc.stderr!,
-            flushMode: .always,
-            lock: flockfile,
-            unlock: funlockfile,
-            write: fwrite,
-            flush: fflush
-        )
-        #endif
-        #elseif canImport(Android)
-        return StdioOutputStream(
-            file: Android.stderr,
-            flushMode: .always,
-            lock: flockfile,
-            unlock: funlockfile,
-            write: fwrite,
-            flush: fflush
-        )
-        #elseif canImport(Musl)
-        return StdioOutputStream(
-            file: Musl.stderr!,
-            flushMode: .always,
-            lock: flockfile,
-            unlock: funlockfile,
-            write: fwrite,
-            flush: fflush
-        )
-        #elseif canImport(WASILibc)
-        // no file locking on WASI
-        return StdioOutputStream(
-            file: WASILibc.stderr!,
-            flushMode: .always,
-            lock: nil,
-            unlock: nil,
-            write: fwrite,
-            flush: fflush
-        )
-        #else
-        #error("Unsupported runtime")
-        #endif
-    }()
-
-    internal static let stdout: StdioOutputStream = {
-        #if canImport(Darwin)
-        return StdioOutputStream(
-            file: Darwin.stdout,
-            flushMode: .always,
-            lock: flockfile,
-            unlock: funlockfile,
-            write: fwrite,
-            flush: fflush
-        )
-        #elseif os(Windows)
-        return StdioOutputStream(
-            file: CRT.stdout,
-            flushMode: .always,
-            lock: _lock_file,
-            unlock: _unlock_file,
-            write: fwrite,
-            flush: fflush
-        )
-        #elseif canImport(Glibc)
-        #if os(FreeBSD) || os(OpenBSD)
-        return StdioOutputStream(
-            file: Glibc.stdout,
-            flushMode: .always,
-            lock: flockfile,
-            unlock: funlockfile,
-            write: fwrite,
-            flush: fflush
-        )
-        #else
-        return StdioOutputStream(
-            file: Glibc.stdout!,
-            flushMode: .always,
-            lock: flockfile,
-            unlock: funlockfile,
-            write: fwrite,
-            flush: fflush
-        )
-        #endif
-        #elseif canImport(Android)
-        return StdioOutputStream(
-            file: Android.stdout,
-            flushMode: .always,
-            lock: flockfile,
-            unlock: funlockfile,
-            write: fwrite,
-            flush: fflush
-        )
-        #elseif canImport(Musl)
-        return StdioOutputStream(
-            file: Musl.stdout!,
-            flushMode: .always,
-            lock: flockfile,
-            unlock: funlockfile,
-            write: fwrite,
-            flush: fflush
-        )
-        #elseif canImport(WASILibc)
-        // no file locking on WASI
-        return StdioOutputStream(
-            file: WASILibc.stdout!,
-            flushMode: .always,
-            lock: nil,
-            unlock: nil,
-            write: fwrite,
-            flush: fflush
-        )
-        #else
-        #error("Unsupported runtime")
-        #endif
-    }()
-
-    /// Defines the flushing strategy for the underlying stream.
-    internal enum FlushMode {
-        case undefined
-        case always
-    }
-}
-
-/// Stream log handler presents log messages to STDERR or STDOUT.
-///
-/// This is a simple implementation of `LogHandler` that directs
-/// `Logger` output to either `stderr` or `stdout` via the factory methods.
-///
-/// Metadata is merged in the following order:
-/// 1. Metadata set on the log handler itself is used as the base metadata.
-/// 2. The handler's ``metadataProvider`` is invoked, overriding any existing keys.
-/// 3. The per-log-statement metadata is merged, overriding any previously set keys.
-public struct StreamLogHandler: LogHandler {
-    internal typealias _SendableTextOutputStream = TextOutputStream & Sendable
-
-    /// Creates a stream log handler that directs its output to STDOUT.
-    public static func standardOutput(label: String) -> StreamLogHandler {
-        StreamLogHandler(
-            label: label,
-            stream: StdioOutputStream.stdout,
-            metadataProvider: LoggingSystem.metadataProvider
-        )
-    }
-
-    /// Creates a stream log handler that directs its output to STDOUT using the metadata provider you provide.
-    public static func standardOutput(label: String, metadataProvider: Logger.MetadataProvider?) -> StreamLogHandler {
-        StreamLogHandler(label: label, stream: StdioOutputStream.stdout, metadataProvider: metadataProvider)
-    }
-
-    /// Creates a stream log handler that directs its output to STDERR.
-    public static func standardError(label: String) -> StreamLogHandler {
-        StreamLogHandler(
-            label: label,
-            stream: StdioOutputStream.stderr,
-            metadataProvider: LoggingSystem.metadataProvider
-        )
-    }
-
-    /// Creates a stream log handler that directs its output to STDERR using the metadata provider you provide.
-    public static func standardError(label: String, metadataProvider: Logger.MetadataProvider?) -> StreamLogHandler {
-        StreamLogHandler(label: label, stream: StdioOutputStream.stderr, metadataProvider: metadataProvider)
-    }
-
-    private let stream: any _SendableTextOutputStream
-    private let label: String
-
-    /// Get the log level configured for this `Logger`.
-    ///
-    /// > Note: Changing the log level threshold for a logger only affects the instance of the `Logger` where you change it.
-    /// > It is acceptable for logging backends to have some form of global log level override
-    /// > that affects multiple or even all loggers. This means a change in `logLevel` to one `Logger` might in
-    /// > certain cases have no effect.
-    public var logLevel: Logger.Level = .info
-
-    /// The metadata provider.
-    public var metadataProvider: Logger.MetadataProvider?
-
-    private var prettyMetadata: String?
-    /// Get or set the entire metadata storage as a dictionary.
-    public var metadata = Logger.Metadata() {
-        didSet {
-            self.prettyMetadata = self.prettify(self.metadata)
-        }
-    }
-
-    /// Add, change, or remove a logging metadata item.
-    ///
-    /// > Note: Changing the logging metadata only affects the instance of the `Logger` where you change it.
-    public subscript(metadataKey metadataKey: String) -> Logger.Metadata.Value? {
-        get {
-            self.metadata[metadataKey]
-        }
-        set {
-            self.metadata[metadataKey] = newValue
-        }
-    }
-
-    // internal for testing only
-    internal init(label: String, stream: any _SendableTextOutputStream) {
-        self.init(label: label, stream: stream, metadataProvider: LoggingSystem.metadataProvider)
-    }
-
-    // internal for testing only
-    internal init(label: String, stream: any _SendableTextOutputStream, metadataProvider: Logger.MetadataProvider?) {
-        self.label = label
-        self.stream = stream
-        self.metadataProvider = metadataProvider
-    }
-
-    /// Log a message using the log level and source that you provide.
-    ///
-    /// - parameters:
-    ///    - event: The log event containing the level, message, metadata, and source location.
-    public func log(event: LogEvent) {
-        let effectiveMetadata = StreamLogHandler.prepareMetadata(
-            base: self.metadata,
-            provider: self.metadataProvider,
-            explicit: event.metadata,
-            error: event.error,
-        )
-
-        let prettyMetadata: String?
-        if let effectiveMetadata = effectiveMetadata {
-            prettyMetadata = self.prettify(effectiveMetadata)
-        } else {
-            prettyMetadata = self.prettyMetadata
-        }
-
-        var stream = self.stream
-        stream.write(
-            "\(self.timestamp()) \(event.level)\(self.label.isEmpty ? "" : " ")\(self.label):\(prettyMetadata.map { " \($0)" } ?? "") [\(event.source)] \(event.message)\n"
-        )
-    }
-
-    @available(*, deprecated, renamed: "log(event:)")
-    public func log(
-        level: Logger.Level,
-        message: Logger.Message,
-        metadata: Logger.Metadata?,
-        source: String,
-        file: String,
-        function: String,
-        line: UInt
-    ) {
-        self.log(
-            event: LogEvent(
-                level: level,
-                message: message,
-                metadata: metadata,
-                source: source,
-                file: file,
-                function: function,
-                line: line
-            )
-        )
-    }
-
-    internal static func prepareMetadata(
-        base: Logger.Metadata,
-        provider: Logger.MetadataProvider?,
-        explicit: Logger.Metadata?,
-        error: (any Error)?,
-    ) -> Logger.Metadata? {
-        var metadata = base
-
-        let provided = provider?.get() ?? [:]
-
-        guard !provided.isEmpty || !((explicit ?? [:]).isEmpty) || error != nil else {
-            // all per-log-statement values are empty
-            return nil
-        }
-
-        if !provided.isEmpty {
-            metadata.merge(provided, uniquingKeysWith: { _, provided in provided })
-        }
-
-        if let explicit = explicit, !explicit.isEmpty {
-            metadata.merge(explicit, uniquingKeysWith: { _, explicit in explicit })
-        }
-
-        if let error {
-            metadata["error.message"] = "\(error)"
-            metadata["error.type"] = "\(String(reflecting: type(of: error)))"
-        }
-
-        return metadata
-    }
-
-    private func prettify(_ metadata: Logger.Metadata) -> String? {
-        if metadata.isEmpty {
-            return nil
-        } else {
-            return metadata.lazy.sorted(by: { $0.key < $1.key }).map { "\($0)=\($1)" }.joined(separator: " ")
-        }
-    }
-
-    private func timestamp() -> String {
-        var buffer = [Int8](repeating: 0, count: 255)
-        #if os(Windows)
-        var timestamp = __time64_t()
-        _ = _time64(&timestamp)
-
-        var localTime = tm()
-        _ = _localtime64_s(&localTime, &timestamp)
-
-        _ = strftime(&buffer, buffer.count, "%Y-%m-%dT%H:%M:%S%z", &localTime)
-        #else
-        var timestamp = time(nil)
-        guard let localTime = localtime(&timestamp) else {
-            return "<unknown>"
-        }
-        strftime(&buffer, buffer.count, "%Y-%m-%dT%H:%M:%S%z", localTime)
-        #endif
-        return buffer.withUnsafeBufferPointer {
-            $0.withMemoryRebound(to: CChar.self) {
-                String(cString: $0.baseAddress!)
-            }
-        }
-    }
-}
-
-/// A no-operation log handler, used when no logging is required
-public struct SwiftLogNoOpLogHandler: LogHandler {
-    /// Creates a no-op log handler.
-    public init() {}
-
-    /// Creates a no-op log handler.
-    public init(_: String) {}
-
-    /// A proxy that discards every log event it receives.
-    ///
-    /// - parameters:
-    ///    - event: The log event to discard.
-    @inlinable public func log(event: LogEvent) {}
-
-    @available(*, deprecated, renamed: "log(event:)")
-    @inlinable public func log(
-        level: Logger.Level,
-        message: Logger.Message,
-        metadata: Logger.Metadata?,
-        source: String,
-        file: String,
-        function: String,
-        line: UInt
-    ) {}
-
-    @available(*, deprecated, renamed: "log(event:)")
-    @inlinable public func log(
-        level: Logging.Logger.Level,
-        message: Logging.Logger.Message,
-        metadata: Logging.Logger.Metadata?,
-        file: String,
-        function: String,
-        line: UInt
-    ) {}
-
-    /// Add, change, or remove a logging metadata item.
-    ///
-    /// > Note: Changing the logging metadata only affects the instance of the `Logger` where you change it.
-    @inlinable public subscript(metadataKey _: String) -> Logger.Metadata.Value? {
-        get {
-            nil
-        }
-        set {}
-    }
-
-    /// Get or set the entire metadata storage as a dictionary.
-    @inlinable public var metadata: Logger.Metadata {
-        get {
-            [:]
-        }
-        set {}
-    }
-
-    /// Get or set the log level configured for this `Logger`.
-    ///
-    /// > Note: Changing the log level threshold for a logger only affects the instance of the `Logger` where you change it.
-    /// > It is acceptable for logging backends to have some form of global log level override
-    /// > that affects multiple or even all loggers. This means a change in `logLevel` to one `Logger` might in
-    /// > certain cases have no effect.
-    @inlinable public var logLevel: Logger.Level {
-        get {
-            .critical
-        }
-        set {}
-    }
-}
-
 extension Logger {
     @inlinable
     internal static func currentModule(filePath: String = #file) -> String {
@@ -2318,24 +1435,6 @@ extension Logger.MetadataValue: ExpressibleByArrayLiteral {
         self = .array(elements)
     }
 }
-
-// MARK: - Debug only warnings
-
-#if DEBUG
-/// Contains state to manage all kinds of "warn only once" warnings which the logging system may want to issue.
-private final class WarnOnceBox: @unchecked Sendable {
-    private let lock: Lock = Lock()
-    private var warnOnceLogHandlerNotSupportedMetadataProviderPerType = Set<ObjectIdentifier>()
-
-    func warnOnceLogHandlerNotSupportedMetadataProvider<Handler: LogHandler>(type: Handler.Type) -> Bool {
-        self.lock.withLock {
-            let id = ObjectIdentifier(type)
-            let (inserted, _) = warnOnceLogHandlerNotSupportedMetadataProviderPerType.insert(id)
-            return inserted  // warn about this handler type, it is the first time we encountered it
-        }
-    }
-}
-#endif
 
 // MARK: - Sendable support helpers
